chore(component): Rename LoadingSpinner -> StateSplash
prep for refactoring

ref: #232 #230
