+++ src/components/StateSplash.tsx
